add EnumVariantName
