+++ libs/gesha-core/src/renderer/rust_type/render_media_type.rs
@@ -1,15 +1,15 @@
 use crate::render;
 use crate::renderer::Result;
-use crate::targets::rust_type::MediaTypeDef;
+use crate::targets::rust_type::{EnumVariantName, MediaTypeDef};
 use std::io::Write;
 
 pub fn render_media_type<W: Write>(mut write: W, x: MediaTypeDef) -> Result<()> {
     println!("media_type: {:#?}", x);
     let variants = x
         .translator
-        .iter()
-        .map(|(&k, _)| k)
-        .collect::<Vec<&str>>()
+        .into_iter()
+        .map(|(k, _)| k)
+        .collect::<Vec<EnumVariantName>>()
         .join(",");
 
     render! { write =>
