+++ libs/gesha-core/src/targets/rust_type/enum_variant.rs
@@ -1,5 +1,4 @@
-use crate::targets::rust_type::DataType;
-use heck::ToUpperCamelCase;
+use crate::targets::rust_type::{DataType, EnumVariantName};
 use std::fmt::{Display, Formatter};
 
 /// rf. https://doc.rust-lang.org/reference/items/enumerations.html
@@ -38,31 +37,6 @@ impl EnumVariant {
 pub enum EnumCase {
     Unit,
     Tuple(Vec<DataType>),
-}
-
-#[derive(Clone, Debug)]
-pub struct EnumVariantName(String);
-
-impl EnumVariantName {
-    pub fn new<A: AsRef<str>>(x: A) -> Self {
-        // TODO: replace x with Rust compatible chars if illegal chars are included
-        Self(x.as_ref().to_upper_camel_case())
-    }
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl Display for EnumVariantName {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        Display::fmt(&self.0, f)
-    }
-}
-
-impl From<EnumVariantName> for String {
-    fn from(this: EnumVariantName) -> Self {
-        this.0
-    }
 }
 
 #[derive(Clone, Debug)]
